Added "getting started" page and stubs for two others
This update adds the getting started page along with the JS code to provide the "workflow" functionality. The nav was reorganized to support multiple content pages and a unified sidebar. JS code was improved with better variable scoping and the work needed for the getting started page and to inject the site email address from the config.